requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.1
python-dotenv==1.0.0
stripe==7.9.0
email-validator==2.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
